Issue #88: ANSI: Replace "log" file with logging module; fix encoding exception.
Previously, when an unexpected character was received by the ANSI FSM,
it would be logged to a file called "log" in the current directory.
As documented in issue #88, if the unexpected character was non-ASCII,
an encoding exception would be reported when this was attempted.

This fix causes the logging module to be used instead of the "log"
file.  This allows applications to control where these messages are
written.  By default - i.e. if the application does not configure the
logging module - these messages will now be discarded.

The message written to the log now includes the repr() of the
unexpected character rather than the character itself, avoiding
exceptions if the character was non-ASCII.  The message is also
improved to be clearer.