Create branch and set version number for 4.0-pre-dcoshea release(s).
As I will need to create one or more releases for my own use, this
commit creates a branch for these and updates the version number for
the first of them.  The version number includes "4.0-pre" as I gather
that the contents are more like 4.0 than 3.3 given the required Python
versions, but note that upstream has yet to release any code with the
version number 4.0, so there is probably quite a bit lacking too.